reorganize cli code
