+++ cowait/cli/app/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa: 401
+from .root import new_cli_app
